# specs/001-controle-financeiro/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Controle Financeiro Pessoal API
  version: 0.1.0
  description: Contrato compartilhado. Mudanças de campo exigem atualizar este arquivo no mesmo PR.
servers:
  - url: http://localhost:8000/api
    description: FastAPI local (backend/)
tags:
  - name: Categorias
  - name: Transacoes
  - name: Relatorios
  - name: Exportacao
  - name: Auth

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Categoria:
      type: object
      required: [id, nome, tipo, cor, icone, padrao]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        tipo: { type: string, enum: [receita, despesa, ambos] }
        cor: { type: string, example: "#22C55E" }
        icone: { type: string, example: utensils }
        padrao: { type: boolean }
    CategoriaInput:
      type: object
      required: [nome, tipo, cor, icone]
      properties:
        nome: { type: string, minLength: 2, maxLength: 40 }
        tipo: { type: string, enum: [receita, despesa, ambos] }
        cor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        icone: { type: string }
    Transacao:
      type: object
      required: [id, tipo, valor, descricao, data, categoriaId]
      properties:
        id: { type: string, format: uuid }
        tipo: { type: string, enum: [receita, despesa] }
        valor: { type: string, description: Decimal como string, example: "150.90" }
        descricao: { type: string }
        data: { type: string, format: date }
        categoriaId: { type: string, format: uuid }
        categoria: { $ref: "#/components/schemas/Categoria" }
    TransacaoInput:
      type: object
      required: [tipo, valor, descricao, data, categoriaId]
      properties:
        tipo: { type: string, enum: [receita, despesa] }
        valor: { type: number, exclusiveMinimum: 0 }
        descricao: { type: string, minLength: 1, maxLength: 120 }
        data: { type: string, format: date }
        categoriaId: { type: string, format: uuid }
    PaginaTransacoes:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transacao" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    ResumoMensal:
      type: object
      required: [ano, mes, totalReceitas, totalDespesas, saldo]
      properties:
        ano: { type: integer }
        mes: { type: integer, minimum: 1, maximum: 12 }
        totalReceitas: { type: string, example: "5000.00" }
        totalDespesas: { type: string, example: "3200.50" }
        saldo: { type: string, example: "1799.50" }
    FatiaPizza:
      type: object
      required: [categoriaId, nome, cor, total]
      properties:
        categoriaId: { type: string, format: uuid }
        nome: { type: string }
        cor: { type: string }
        total: { type: string }
    PontoEvolucao:
      type: object
      required: [ano, mes, totalReceitas, totalDespesas]
      properties:
        ano: { type: integer }
        mes: { type: integer }
        totalReceitas: { type: string }
        totalDespesas: { type: string }
    Comparativo:
      type: object
      required: [atual, anterior, variacaoReceitas, variacaoDespesas, variacaoSaldo]
      properties:
        atual: { $ref: "#/components/schemas/ResumoMensal" }
        anterior: { $ref: "#/components/schemas/ResumoMensal" }
        variacaoReceitas: { type: string }
        variacaoDespesas: { type: string }
        variacaoSaldo: { type: string }
    UsuarioPublico:
      type: object
      required: [id, nome, email]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        email: { type: string, format: email }
    RegistroInput:
      type: object
      required: [nome, email, senha]
      properties:
        nome: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        senha: { type: string, minLength: 8, maxLength: 128 }
    LoginInput:
      type: object
      required: [email, senha]
      properties:
        email: { type: string, format: email }
        senha: { type: string }
    Sessao:
      type: object
      required: [accessToken, tokenType, usuario]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, example: bearer }
        usuario: { $ref: "#/components/schemas/UsuarioPublico" }

paths:
  /categorias:
    get:
      tags: [Categorias]
      summary: Lista categorias do usuário
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Categoria" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Categorias]
      summary: Cria categoria customizada
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoriaInput" }
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Categoria" }
        "400":
          description: Validação
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /categorias/{id}:
    patch:
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoriaInput" }
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Categoria" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { description: Não encontrada }
    delete:
      tags: [Categorias]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Removida }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Padrão ou em uso
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /transacoes:
    get:
      tags: [Transacoes]
      summary: Lista paginada de transações
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginaTransacoes" }
        "400":
          description: Parâmetros de paginação inválidos.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Transacoes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransacaoInput" }
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transacao" }
        "400": { description: Validação }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /transacoes/{id}:
    patch:
      tags: [Transacoes]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransacaoInput" }
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transacao" }
        "400":
          description: Payload, UUID ou categoria inválidos, ou tipo incompatível com a categoria.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Transação inexistente ou pertencente a outro usuário.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Transacoes]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Removida }
        "400":
          description: UUID inválido.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Transação inexistente ou pertencente a outro usuário.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /relatorios/resumo:
    get:
      tags: [Relatorios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          required: true
          schema: { type: integer }
        - in: query
          name: mes
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResumoMensal" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /relatorios/pizza:
    get:
      tags: [Relatorios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          required: true
          schema: { type: integer }
        - in: query
          name: mes
          required: true
          schema: { type: integer }
        - in: query
          name: tipo
          schema: { type: string, enum: [despesa, receita], default: despesa }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FatiaPizza" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /relatorios/evolucao:
    get:
      tags: [Relatorios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: meses
          schema: { type: integer, default: 6, maximum: 24 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PontoEvolucao" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /relatorios/comparativo:
    get:
      tags: [Relatorios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: ano
          required: true
          schema: { type: integer }
        - in: query
          name: mes
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comparativo" }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /transacoes/export:
    get:
      tags: [Exportacao]
      summary: CSV da listagem (mesmos filtros de GET /transacoes)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: text/csv
          content:
            text/csv:
              schema: { type: string }
        "401":
          description: Sessão ausente ou inválida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/register:
    post:
      tags: [Auth]
      description: Registro no FastAPI (`POST /api/auth/register`). Cria a conta, categorias padrão e inicia a sessão JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegistroInput" }
      responses:
        "201":
          description: Conta criada e sessão iniciada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sessao" }
        "400":
          description: Validação
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: E-mail já existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Sessão iniciada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sessao" }
        "401":
          description: E-mail ou senha inválidos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Sessão encerrada no cliente }
        "401":
          description: Sessão ausente ou inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Usuário da sessão
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioPublico" }
        "401":
          description: Sessão ausente ou inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }